requirements.txt
openpyxl
smtplib3
Flask
Werkzeug
google-generativeai
Gunicorn